Add files via upload
We use Data Analytics in everyday life without even knowing it. The task is to give an example (s) of a real-life situation where we use Data Analytics and link it with the data analytics process. You can prepare a PPT/PDF on a real-life scenario explaining it with the above process (Plan, Prepare, Process, Analyze, Share, Act) and submit it as part of this task.